requirements.txt
fastapi
pydantic
tavily-python
google-genai